Extension Modifications, Precedence Alterations & Clean-up - Internalizing a number of the extension methods that were never intended to to be exposed publicly. This is a potentially breaking change, though it should be unlikely in most instances. - Modifying precedence attribute to provide greater extensibility by converting the precedence value to a double in order to make way for a broader range of precedents. Also makes it easier to inject new precedence's between those that already exist. - Removing PathHelper's which we ported but never needed. This also is a potentially breaking change, though it should be unlikely in most instances. - Cleaning up erant references in the test project.

+++ Functionless/Injection/Module.cs
@@ -27,7 +27,7 @@ namespace Functionless.Injection
                 assemblies
                     .SelectMany(s => s.GetTypes())
                     .Where(p => p.IsAssignableTo<IInterceptor>())
-                    .OrderBy(t => t.GetCustomAttribute<PrecedenceAttribute>()?.Value ?? int.MaxValue)
+                    .OrderBy(t => t.GetCustomAttribute<PrecedenceAttribute>()?.Value ?? double.MaxValue)
                     .ToArray();
 
             builder.RegisterTypes(interceptorTypes).AsSelf().SingleInstance();
